cli: rename cmd/app package to cmd/horusec
Rename the package cmd/app to cmd/horusec to users be able to install
the binary using go install github.com/ZupIT/horusec/cmd/horusec@latest
