+++ cmd/horusec/start/start_test.go
@@ -370,7 +370,7 @@ func TestStartCommand_Execute(t *testing.T) {
 		assert.Contains(t, output, "FOLDER BEFORE THE ANALYSIS FINISH! Don’t worry, we’ll remove it after the analysis ends automatically! Project sent to folder in location: ")
 		assert.Contains(t, output, "Horusec will return a timeout after 600 seconds. This time can be customized in the cli settings.")
 		assert.Contains(t, output, "{HORUSEC_CLI} Writing output JSON to file in the path:")
-		assert.Contains(t, output, "cmd/app/start/tmp-json.json")
+		assert.Contains(t, output, "cmd/horusec/start/tmp-json.json")
 		assert.Contains(t, output, "{HORUSEC_CLI} No authorization token was found, your code it is not going to be sent to horusec. Please enter a token with the -a flag to configure and save your analysis")
 		assert.Contains(t, output, "YOUR ANALYSIS HAD FINISHED WITHOUT ANY VULNERABILITY!")
 		assert.Contains(t, output, "{HORUSEC_CLI} Horusec not show info vulnerabilities in this analysis")
@@ -530,7 +530,7 @@ func TestStartCommand_Execute(t *testing.T) {
 		assert.Contains(t, output, "FOLDER BEFORE THE ANALYSIS FINISH! Don’t worry, we’ll remove it after the analysis ends automatically! Project sent to folder in location: ")
 		assert.Contains(t, output, "Horusec will return a timeout after 600 seconds. This time can be customized in the cli settings.")
 		assert.Contains(t, output, "{HORUSEC_CLI} Writing output JSON to file in the path:")
-		assert.Contains(t, output, "cmd/app/start/tmp-sonarqube.json")
+		assert.Contains(t, output, "cmd/horusec/start/tmp-sonarqube.json")
 		assert.Contains(t, output, "{HORUSEC_CLI} No authorization token was found, your code it is not going to be sent to horusec. Please enter a token with the -a flag to configure and save your analysis")
 		assert.Contains(t, output, "YOUR ANALYSIS HAD FINISHED WITHOUT ANY VULNERABILITY!")
 		assert.Contains(t, output, "{HORUSEC_CLI} Horusec not show info vulnerabilities in this analysis")
