+++ goreleaser.yml
@@ -16,7 +16,7 @@ project_name: horusec
 builds:
   - id: horusec
     env: [ CGO_ENABLED=0 ]
-    main: ./cmd/app/main.go
+    main: ./cmd/horusec/main.go
     ldflags:
       - -X github.com/ZupIT/horusec/config.version=v{{ .Version }}
       - -X github.com/ZupIT/horusec/config.commit={{ .Commit }}
